backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
sqlalchemy
argon2-cffi
python-jose
python-multipart
requests
torch
sentence-transformers
